Extracts hydration functionality into a `common` package
